Split card to Header, Body and Footer
