+++ app/javascript/src/components/Notes/Card/Body.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Body = ({ description }) => (
+  <div className="font-normal text-gray-500 text-sm border-b pb-2">
+    {description}
+  </div>
+);
+
+export default Body;
